+++ ComparerExtensions/ComparerExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Reflection;
 
 namespace ComparerExtensions
 {
@@ -23,13 +22,13 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             if (comparer == null)
             {
-                throw new ArgumentNullException("comparer");
+                throw new ArgumentNullException(nameof(comparer));
             }
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -46,14 +45,14 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             if (comparison == null)
             {
-                throw new ArgumentNullException("comparison");
+                throw new ArgumentNullException(nameof(comparison));
             }
-            IComparer<T> wrapper = ComparisonWrapper<T>.GetComparer(comparison);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, wrapper);
+            var wrapper = ComparisonWrapper<T>.GetComparer(comparison);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, wrapper);
             return compoundComparer;
         }
 
@@ -71,10 +70,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T> comparer = KeyComparer<T>.OrderBy(keySelector);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -94,10 +93,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T> comparer = KeyComparer<T>.OrderBy(keySelector, keyComparer);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -117,10 +116,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T> comparer = KeyComparer<T>.OrderBy(keySelector, keyComparison);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -138,10 +137,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T> comparer = KeyComparer<T>.OrderByDescending(keySelector);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -161,10 +160,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T> comparer = KeyComparer<T>.OrderByDescending(keySelector, keyComparer);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -186,10 +185,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T> comparer = KeyComparer<T>.OrderByDescending(keySelector, keyComparison);
-            IComparer<T> compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
+            var compoundComparer = CompoundComparer<T>.GetComparer(baseComparer, comparer);
             return compoundComparer;
         }
 
@@ -206,10 +205,10 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             IComparer<T?> uncheckedComparer = new UncheckedNullableComparer<T>(baseComparer);
-            return UnkeyedNullPlacementComparer<T?>.GetComparer(uncheckedComparer, nullsFirst: true);
+            return UnkeyedNullPlacementComparer<T?>.GetComparer(uncheckedComparer, true);
         }
 
         /// <summary>
@@ -223,16 +222,15 @@ namespace ComparerExtensions
         /// <remarks>The returned comparer is guaranteed to never pass a null value to the given base comparer.</remarks>
         public static IComparer<T> NullsFirst<T>(this IComparer<T> baseComparer)
         {
-            if (baseComparer == null)
+            switch (baseComparer)
             {
-                throw new ArgumentNullException("baseComparer");
+                case null:
+                    throw new ArgumentNullException(nameof(baseComparer));
+                case IPrecedenceEnforcer<T> previous:
+                    return previous.CreateUnkeyedComparer(true);
             }
-            IPrecedenceEnforcer<T> previous = baseComparer as IPrecedenceEnforcer<T>;
-            if (previous != null)
-            {
-                return previous.CreateUnkeyedComparer(nullsFirst: true);
-            }
-            return UnkeyedNullPlacementComparer<T>.GetComparer(baseComparer, nullsFirst: true);
+
+            return UnkeyedNullPlacementComparer<T>.GetComparer(baseComparer, true);
         }
 
         /// <summary>
@@ -246,16 +244,15 @@ namespace ComparerExtensions
         /// <remarks>The returned comparer is guaranteed to never pass a null value to the given base comparer.</remarks>
         public static IComparer<T> NullsLast<T>(this IComparer<T> baseComparer)
         {
-            if (baseComparer == null)
+            switch (baseComparer)
             {
-                throw new ArgumentNullException("baseComparer");
+                case null:
+                    throw new ArgumentNullException(nameof(baseComparer));
+                case IPrecedenceEnforcer<T> previous:
+                    return previous.CreateUnkeyedComparer(false);
             }
-            IPrecedenceEnforcer<T> previous = baseComparer as IPrecedenceEnforcer<T>;
-            if (previous != null)
-            {
-                return previous.CreateUnkeyedComparer(nullsFirst: false);
-            }
-            return UnkeyedNullPlacementComparer<T>.GetComparer(baseComparer, nullsFirst: false);
+
+            return UnkeyedNullPlacementComparer<T>.GetComparer(baseComparer, false);
         }
 
         /// <summary>
@@ -274,18 +271,17 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             if (keySelector == null)
             {
-                throw new ArgumentNullException("keySelector");
+                throw new ArgumentNullException(nameof(keySelector));
             }
-            IPrecedenceEnforcer<T> previous = baseComparer as IPrecedenceEnforcer<T>;
-            if (previous != null)
+            if (baseComparer is IPrecedenceEnforcer<T> previous)
             {
-                return previous.CreateKeyedComparer(keySelector, nullsFirst: true);
+                return previous.CreateKeyedComparer(keySelector, true);
             }
-            return KeyedNullPlacementComparer<T, TKey>.GetComparer(baseComparer, keySelector, nullsFirst: true);
+            return KeyedNullPlacementComparer<T, TKey>.GetComparer(baseComparer, keySelector, true);
         }
 
         /// <summary>
@@ -304,18 +300,17 @@ namespace ComparerExtensions
         {
             if (baseComparer == null)
             {
-                throw new ArgumentNullException("baseComparer");
+                throw new ArgumentNullException(nameof(baseComparer));
             }
             if (keySelector == null)
             {
-                throw new ArgumentNullException("keySelector");
+                throw new ArgumentNullException(nameof(keySelector));
             }
-            IPrecedenceEnforcer<T> previous = baseComparer as IPrecedenceEnforcer<T>;
-            if (previous != null)
+            if (baseComparer is IPrecedenceEnforcer<T> previous)
             {
-                return previous.CreateKeyedComparer(keySelector, nullsFirst: false);
+                return previous.CreateKeyedComparer(keySelector, false);
             }
-            return KeyedNullPlacementComparer<T, TKey>.GetComparer(baseComparer, keySelector, nullsFirst: false);
+            return KeyedNullPlacementComparer<T, TKey>.GetComparer(baseComparer, keySelector, false);
         }
 
         /// <summary>
@@ -329,7 +324,7 @@ namespace ComparerExtensions
         {
             if (comparer == null)
             {
-                throw new ArgumentNullException("comparer");
+                throw new ArgumentNullException(nameof(comparer));
             }
             return new ReversedComparer<T>(comparer);
         }
@@ -345,7 +340,7 @@ namespace ComparerExtensions
         {
             if (comparer == null)
             {
-                throw new ArgumentNullException("comparer");
+                throw new ArgumentNullException(nameof(comparer));
             }
             return TypedComparer<T>.GetComparer(comparer);
         }
@@ -360,7 +355,7 @@ namespace ComparerExtensions
         {
             if (comparer == null)
             {
-                throw new ArgumentNullException("comparer");
+                throw new ArgumentNullException(nameof(comparer));
             }
             return UntypedComparer<T>.GetComparer(comparer);
         }
@@ -376,449 +371,9 @@ namespace ComparerExtensions
         {
             if (comparison == null)
             {
-                throw new ArgumentNullException("comparison");
+                throw new ArgumentNullException(nameof(comparison));
             }
             return ComparisonWrapper<T>.GetComparer(comparison);
-        }
-    }
-
-    internal sealed class TypedComparer<T> : IComparer<T>, IComparer
-    {
-        private readonly IComparer _comparer;
-
-        public static IComparer<T> GetComparer(IComparer comparer)
-        {
-            UntypedComparer<T> untyped = comparer as UntypedComparer<T>;
-            if (untyped != null)
-            {
-                return untyped.Comparer;
-            }
-            IComparer<T> typed = comparer as IComparer<T>;
-            if (typed != null)
-            {
-                return typed;
-            }
-            return new TypedComparer<T>(comparer);
-        }
-
-        private TypedComparer(IComparer comparer)
-        {
-            _comparer = comparer;
-        }
-
-        public IComparer Comparer
-        {
-            get { return _comparer; }
-        }
-
-        public int Compare(T x, T y)
-        {
-            return _comparer.Compare(x, y);
-        }
-
-        int IComparer.Compare(object x, object y)
-        {
-            return _comparer.Compare(x, y);
-        }
-    }
-
-    internal sealed class UntypedComparer<T> : IComparer<T>, IComparer
-    {
-        private readonly IComparer<T> _comparer;
-
-        public static IComparer GetComparer(IComparer<T> comparer)
-        {
-            TypedComparer<T> typed = comparer as TypedComparer<T>;
-            if (typed != null)
-            {
-                return typed.Comparer;
-            }
-            IComparer untyped = comparer as IComparer;
-            if (untyped != null)
-            {
-                return untyped;
-            }
-            return new UntypedComparer<T>(comparer);
-        }
-
-        private UntypedComparer(IComparer<T> comparer)
-        {
-            _comparer = comparer;
-        }
-
-        public IComparer<T> Comparer
-        {
-            get { return _comparer; }
-        }
-
-        public int Compare(T x, T y)
-        {
-            return _comparer.Compare(x, y);
-        }
-
-        int IComparer.Compare(object x, object y)
-        {
-            return _comparer.Compare((T)x, (T)y);
-        }
-    }
-
-    internal sealed class ComparisonWrapper<T> : Comparer<T>
-    {
-        private readonly Func<T, T, int> _comparison;
-
-        public static IComparer<T> GetComparer(Func<T, T, int> comparison)
-        {
-            IComparer<T> source = comparison.Target as IComparer<T>;
-            return source ?? new ComparisonWrapper<T>(comparison);
-        }
-
-        private ComparisonWrapper(Func<T, T, int> comparison)
-        {
-            _comparison = comparison;
-        }
-
-        public override int Compare(T x, T y)
-        {
-            return _comparison(x, y);
-        }
-    }
-
-    internal sealed class CompoundComparer<T> : Comparer<T>
-    {
-        private readonly List<IComparer<T>> _comparers;
-
-        public static IComparer<T> GetComparer(IComparer<T> baseComparer, IComparer<T> nextComparer)
-        {
-            // make sure null comparer stays highest precedence
-            IPrecedenceEnforcer<T> nullComparer = baseComparer as IPrecedenceEnforcer<T>;
-            if (nullComparer != null)
-            {
-                return nullComparer.CreateCompoundComparer(nextComparer);
-            }
-            CompoundComparer<T> comparer = new CompoundComparer<T>();
-            comparer.AppendComparison(baseComparer);
-            comparer.AppendComparison(nextComparer);
-            return comparer.Normalize();
-        }
-
-        public CompoundComparer()
-        {
-            _comparers = new List<IComparer<T>>();
-        }
-
-        public void AppendComparison(IComparer<T> comparer)
-        {
-            if (comparer is NullComparer<T>)
-            {
-                return;
-            }
-            CompoundComparer<T> other = comparer as CompoundComparer<T>;
-            if (other != null)
-            {
-                _comparers.AddRange(other._comparers);
-                return;
-            }
-            _comparers.Add(comparer);
-        }
-
-        public override int Compare(T x, T y)
-        {
-            foreach (IComparer<T> comparer in _comparers)
-            {
-                int result = comparer.Compare(x, y);
-                if (result != 0)
-                {
-                    return result;
-                }
-            }
-            return 0;
-        }
-
-        public IComparer<T> Normalize()
-        {
-            if (_comparers.Count == 0)
-            {
-                return NullComparer<T>.Default;
-            }
-            if (_comparers.Count == 1)
-            {
-                return _comparers[0];
-            }
-            return this;
-        }
-    }
-
-    internal sealed class ReversedComparer<T> : Comparer<T>
-    {
-        private readonly IComparer<T> _comparer;
-
-        public ReversedComparer(IComparer<T> comparer)
-        {
-            _comparer = comparer;
-        }
-
-        public override int Compare(T x, T y)
-        {
-            return _comparer.Compare(y, x);
-        }
-    }
-
-    internal sealed class UncheckedNullableComparer<T> : Comparer<T?>
-        where T : struct
-    {
-        private readonly IComparer<T> _comparer;
-
-        public UncheckedNullableComparer(IComparer<T> comparer)
-        {
-            _comparer = comparer;
-        }
-
-        public override int Compare(T? x, T? y)
-        {
-            return _comparer.Compare(x.Value, y.Value);
-        }
-    }
-
-    internal interface IPrecedenceEnforcer<T>
-    {
-        IComparer<T> CreateUnkeyedComparer(bool nullsFirst);
-
-        IComparer<T> CreateKeyedComparer<TKey>(Func<T, TKey> keySelector, bool nullsFirst);
-
-        IComparer<T> CreateCompoundComparer(IComparer<T> comparer);
-    }
-
-    internal abstract class NullPlacementComparer<T, TCompared> : Comparer<T>
-    {
-        protected readonly IComparer<T> Comparer;
-        protected readonly NullFilter<TCompared> NullFilter;
-
-        protected NullPlacementComparer(IComparer<T> comparer, NullFilter<TCompared> filter)
-        {
-            Comparer = comparer;
-            NullFilter = filter;
-        }
-    }
-
-    internal sealed class UnkeyedNullPlacementComparer<T> : NullPlacementComparer<T, T>, IPrecedenceEnforcer<T>
-    {
-        public static IComparer<T> GetComparer(IComparer<T> comparer, bool nullsFirst)
-        {
-            NullFilter<T> filter = NullFilterFactory.GetNullFilter<T>(nullsFirst);
-            if (filter == null)
-            {
-                return comparer;
-            }
-            return new UnkeyedNullPlacementComparer<T>(comparer, filter);
-        }
-
-        internal UnkeyedNullPlacementComparer(IComparer<T> comparer, NullFilter<T> filter)
-            : base(comparer, filter)
-        {
-        }
-
-        public override int Compare(T x, T y)
-        {
-            return NullFilter.Filter(x, y) ?? Comparer.Compare(x, y);
-        }
-
-        public IComparer<T> CreateUnkeyedComparer(bool nullsFirst)
-        {
-            // we're duplicating ourselves
-            if (nullsFirst == NullFilter.NullsFirst)
-            {
-                return this;
-            }
-            // we're replacing ourselves with another top-level comparer with a different sort order
-            return UnkeyedNullPlacementComparer<T>.GetComparer(Comparer, nullsFirst);
-        }
-
-        public IComparer<T> CreateKeyedComparer<TKey>(Func<T, TKey> keySelector, bool nullsFirst)
-        {
-            // we want to take precedence over the new keyed comparer
-            IComparer<T> comparer = KeyedNullPlacementComparer<T, TKey>.GetComparer(Comparer, keySelector, nullsFirst);
-            return UnkeyedNullPlacementComparer<T>.GetComparer(comparer, NullFilter.NullsFirst);
-        }
-
-        public IComparer<T> CreateCompoundComparer(IComparer<T> comparer)
-        {
-            CompoundComparer<T> compoundComparer = new CompoundComparer<T>();
-            compoundComparer.AppendComparison(Comparer);
-            compoundComparer.AppendComparison(comparer);
-            return UnkeyedNullPlacementComparer<T>.GetComparer(compoundComparer.Normalize(), NullFilter.NullsFirst);
-        }
-    }
-
-    internal sealed class KeyedNullPlacementComparer<T, TKey> : NullPlacementComparer<T, TKey>, IPrecedenceEnforcer<T>
-    {
-        private readonly Func<T, TKey> _keySelector;
-
-        public static IComparer<T> GetComparer(IComparer<T> comparer, Func<T, TKey> keySelector, bool nullsFirst)
-        {
-            NullFilter<TKey> filter = NullFilterFactory.GetNullFilter<TKey>(nullsFirst);
-            if (filter == null)
-            {
-                return comparer;
-            }
-            return new KeyedNullPlacementComparer<T, TKey>(comparer, keySelector, filter);
-        }
-
-        internal KeyedNullPlacementComparer(IComparer<T> comparer, Func<T, TKey> keySelector, NullFilter<TKey> filter)
-            : base(comparer, filter)
-        {
-            _keySelector = keySelector;
-        }
-
-        public override int Compare(T x, T y)
-        {
-            TKey xKey = _keySelector(x);
-            TKey yKey = _keySelector(y);
-            return NullFilter.Filter(xKey, yKey) ?? Comparer.Compare(x, y);
-        }
-
-        public IComparer<T> CreateUnkeyedComparer(bool nullsFirst)
-        {
-            // top-level comparers should be evaluated before keyed comparers
-            return UnkeyedNullPlacementComparer<T>.GetComparer(this, nullsFirst);
-        }
-
-        public IComparer<T> CreateKeyedComparer<TOtherKey>(Func<T, TOtherKey> keySelector, bool nullsFirst)
-        {
-            // we don't know if the key selector is identical, so we must wrap ourselves
-            return KeyedNullPlacementComparer<T, TOtherKey>.GetComparer(this, keySelector, nullsFirst);
-        }
-
-        public IComparer<T> CreateCompoundComparer(IComparer<T> comparer)
-        {
-            CompoundComparer<T> compoundComparer = new CompoundComparer<T>();
-            compoundComparer.AppendComparison(Comparer);
-            compoundComparer.AppendComparison(comparer);
-            return KeyedNullPlacementComparer<T, TKey>.GetComparer(compoundComparer.Normalize(), _keySelector, NullFilter.NullsFirst);
-        }
-    }
-
-    internal static class NullFilterFactory
-    {
-        private static readonly Dictionary<Type, object[]> _filterTypeLookup = new Dictionary<Type, object[]>();
-
-        public static NullFilter<T> GetNullFilter<T>(bool nullsFirst)
-        {
-            Type comparedType = typeof(T);
-            object[] filters;
-            if (!_filterTypeLookup.TryGetValue(comparedType, out filters))
-            {
-                filters = getTypeFilters<T>(comparedType, filters);
-                _filterTypeLookup.Add(comparedType, filters);
-            }
-            return (NullFilter<T>)filters[nullsFirst ? 1 : 0];
-        }
-
-        private static object[] getTypeFilters<T>(Type comparedType, object[] filters)
-        {
-            Type underlyingType = Nullable.GetUnderlyingType(comparedType);
-            if (underlyingType != null)
-            {
-                Type genericFilterType = typeof(NullableNullFilter<>);
-                Type filterType = genericFilterType.MakeGenericType(underlyingType);
-                return new object[]
-                { 
-                    Activator.CreateInstance(filterType, new object[] { false }),
-                    Activator.CreateInstance(filterType, new object[] { true }),
-                };
-            }
-            if (comparedType.GetTypeInfo().IsValueType)
-            {
-                return new object[2];
-            }
-            return new object[] 
-            { 
-                new ReferenceNullFilter<T>(false),
-                new ReferenceNullFilter<T>(true),
-            };
-        }
-    }
-
-    internal abstract class NullFilter<T>
-    {
-        private readonly bool _nullsFirst;
-
-        protected NullFilter(bool nullsFirst)
-        {
-            _nullsFirst = nullsFirst;
-        }
-
-        public bool NullsFirst
-        {
-            get { return _nullsFirst; }
-        }
-
-        public abstract int? Filter(T x, T y);
-    }
-
-    internal sealed class NullableNullFilter<T> : NullFilter<T?>
-        where T : struct
-    {
-        public NullableNullFilter(bool nullsFirst)
-            : base(nullsFirst)
-        {
-        }
-
-        public override int? Filter(T? x, T? y)
-        {
-            if (x.HasValue)
-            {
-                if (y.HasValue)
-                {
-                    return null;
-                }
-                if (NullsFirst)
-                {
-                    return 1;
-                }
-                return -1;
-            }
-            if (y.HasValue)
-            {
-                if (NullsFirst)
-                {
-                    return -1;
-                }
-                return 1;
-            }
-            return 0;
-        }
-    }
-
-    internal sealed class ReferenceNullFilter<T> : NullFilter<T>
-    {
-        public ReferenceNullFilter(bool nullsFirst)
-            : base(nullsFirst)
-        {
-        }
-
-        public override int? Filter(T x, T y)
-        {
-            if (x == null)
-            {
-                if (y == null)
-                {
-                    return 0;
-                }
-                if (NullsFirst)
-                {
-                    return -1;
-                }
-                return 1;
-            }
-            if (y == null)
-            {
-                if (NullsFirst)
-                {
-                    return 1;
-                }
-                return -1;
-            }
-            return null;
         }
     }
 }
